Merge upstream changes  (#120)
* Fix exponentialMovingAverage panic

The panic was happening when the window size was greater than
the length of the values. Now if that happens no panics occur and
the average is returned. This matches graphiteweb.

* Add support for Graphite web toLowerCase function

* Add tests for toLowerCase function

* Update glue.go

* Update COMPATIBILITY.md to show support of toLowerCase

* Add support for Graphite web toUpperCase function

* Add tests for toUpperCase function

* Update glue.go

* Update COMPATIBILITY.md to include toUpperCase function

* Add support for Graphite web sumSeriesLists function

* Add tests for sumSeriesLists

* Update glue.go

* Update COMPATIBILITY.md to include sumSeriesLists

* Fix problems that I accidentally introduced while trying to resolve merge conflict

* Another attempt to fix previous problem

 Apparantly there is a directory that gen.go wasn't expecting and
 `go generate` in expr/functions produced `glue.go` file that didn't compile

 fix both generator and glue

* Adds Drone and Github workflow for Grafana's fork

This commit adds the files that are specific to Grafana's fork
that are not intended to be included in the upstream. This
files are already present in the commit history (they were deleted),
but this is an attempt of a fresh start on top of the upstream's `main`.
At the moment, this should the only commit we expect to skip when
sending our changes.

Co-authored-by: Nicolás Pazos <npazosmendez@gmail.com>
Co-authored-by: Vladimir Smirnov <civiloid@google.com>